refactor(playback::playlist): move save to file & load from file to own module
