+++ playback/src/playlist/mod.rs
@@ -5,8 +5,6 @@ use rand::RngExt;
 use rand::seq::SliceRandom;
 use std::error::Error;
 use std::fmt::{Display, Write as _};
-use std::fs::File;
-use std::io::{BufRead, BufReader, BufWriter, Write};
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use std::time::{SystemTime, UNIX_EPOCH};
@@ -35,6 +33,7 @@ use crate::SharedPlaylist;
 use crate::StreamTX;
 use crate::playlist::sorting::{ScoredTrack, score_track, sort_scored};
 
+mod save_load;
 mod sorting;
 
 #[derive(Debug)]
@@ -182,93 +181,13 @@ impl Playlist {
         self.next_track_index.and_then(|v| self.tracks.get(v))
     }
 
-    /// Load the playlist from the file.
-    ///
-    /// Path in `$config$/playlist.log`.
-    ///
-    /// Returns `(Position, Tracks[])`.
-    ///
-    /// # Errors
-    /// - When the playlist path is not write-able
-    /// - When podcasts cannot be loaded
-    pub fn load() -> Result<(usize, Vec<Track>)> {
-        let path = get_playlist_path()?;
-
-        let Ok(file) = File::open(&path) else {
-            // new file, nothing to parse from it
-            File::create(&path)?;
-
-            return Ok((0, Vec::new()));
-        };
-
-        let reader = BufReader::new(file);
-        let mut lines = reader.lines();
-
-        let mut current_track_index = 0;
-        if let Some(line) = lines.next() {
-            let index_line = line?;
-            if let Ok(index) = index_line.trim().parse() {
-                current_track_index = index;
-            }
-        } else {
-            // empty file, nothing to parse from it
-            return Ok((0, Vec::new()));
-        }
-
-        let mut playlist_items = Vec::new();
-        let db_path = get_app_config_path()?;
-        let db_podcast = DBPod::new(&db_path)?;
-        let podcasts = db_podcast
-            .get_podcasts()
-            .with_context(|| "failed to get podcasts from db.")?;
-        for line in lines {
-            let line = line?;
-
-            let trimmed_line = line.trim();
-
-            // skip empty lines without trying to process them
-            // skip lines that are comments (m3u-like)
-            if trimmed_line.is_empty() || trimmed_line.starts_with('#') {
-                continue;
-            }
-
-            if line.starts_with("http") {
-                let mut is_podcast = false;
-                'outer: for pod in &podcasts {
-                    for ep in &pod.episodes {
-                        if ep.url == line.as_str() {
-                            is_podcast = true;
-                            let track = Track::from_podcast_episode(ep);
-                            playlist_items.push(track);
-                            break 'outer;
-                        }
-                    }
-                }
-                if !is_podcast {
-                    let track = Track::new_radio(&line);
-                    playlist_items.push(track);
-                }
-                continue;
-            }
-            if let Ok(track) = Track::read_track_from_path(&line) {
-                playlist_items.push(track);
-            }
-        }
-
-        // protect against the listed index in the playlist file not matching the elements in the playlist
-        // for example lets say it has "100", but there are only 2 elements in the playlist
-        let current_track_index = current_track_index.min(playlist_items.len().saturating_sub(1));
-
-        Ok((current_track_index, playlist_items))
-    }
-
     /// Run [`load`](Self::load), but also apply the values directly to the current instance.
     ///
     /// # Errors
     ///
     /// See [`load`](Self::load)
     pub fn load_apply(&mut self) -> Result<()> {
-        let (current_track_index, tracks) = Self::load()?;
+        let (current_track_index, tracks) = save_load::load()?;
         self.current_track_index = current_track_index;
         self.tracks = tracks;
         self.is_modified = false;
@@ -331,7 +250,7 @@ impl Playlist {
     ///
     /// See [`Self::load`]
     pub fn reload_tracks(&mut self) -> Result<()> {
-        let (current_track_index, tracks) = Self::load()?;
+        let (current_track_index, tracks) = save_load::load()?;
         self.tracks = tracks;
         self.current_track_index = current_track_index;
         self.is_modified = false;
@@ -347,29 +266,7 @@ impl Playlist {
     ///
     /// Errors could happen when writing files
     pub fn save(&mut self) -> Result<()> {
-        let path = get_playlist_path()?;
-
-        let file = File::create(&path)?;
-
-        // If the playlist is empty, truncate the file, but dont write anything else (like a index number)
-        if self.is_empty() {
-            self.is_modified = false;
-            return Ok(());
-        }
-
-        let mut writer = BufWriter::new(file);
-        writer.write_all(self.current_track_index.to_string().as_bytes())?;
-        writer.write_all(b"\n")?;
-        for track in &self.tracks {
-            let id = match track.inner() {
-                MediaTypes::Track(track_data) => track_data.path().to_string_lossy(),
-                MediaTypes::Radio(radio_track_data) => radio_track_data.url().into(),
-                MediaTypes::Podcast(podcast_track_data) => podcast_track_data.url().into(),
-            };
-            writeln!(writer, "{id}")?;
-        }
-
-        writer.flush()?;
+        save_load::save(self)?;
         self.is_modified = false;
 
         Ok(())
